Move main to cmd package
As this project will have many small programs in the future, we need to
refactory now so the tool follows the common directory structure from other Go
projects.

See #52
