+++ toglacier.go
@@ -1,311 +1,30 @@
-package main
+package toglacier
 
 import (
 	"context"
 	"fmt"
-	"io"
 	"net/smtp"
 	"os"
-	"os/signal"
 	"sort"
 	"strings"
-	"syscall"
 	"time"
 
-	"github.com/Sirupsen/logrus"
-	"github.com/jasonlvhit/gocron"
 	"github.com/pkg/errors"
 	"github.com/rafaeljusto/toglacier/internal/archive"
 	"github.com/rafaeljusto/toglacier/internal/cloud"
-	"github.com/rafaeljusto/toglacier/internal/config"
 	"github.com/rafaeljusto/toglacier/internal/log"
 	"github.com/rafaeljusto/toglacier/internal/report"
 	"github.com/rafaeljusto/toglacier/internal/storage"
-	"github.com/urfave/cli"
 )
-
-func main() {
-	var toGlacier ToGlacier
-	var logger *logrus.Logger
-	var logFile *os.File
-	defer logFile.Close()
-
-	// ctx is used to abort long transactions, such as big files uploads or
-	// inventories
-	ctx := context.Background()
-	ctx, cancel := context.WithCancel(ctx)
-	var cancelFunc func()
-
-	app := cli.NewApp()
-	app.Name = "toglacier"
-	app.Usage = "Send data to AWS Glacier service"
-	app.Version = "2.0.0"
-	app.Authors = []cli.Author{
-		{
-			Name:  "Rafael Dantas Justo",
-			Email: "adm@rafael.net.br",
-		},
-	}
-	app.Flags = []cli.Flag{
-		cli.StringFlag{
-			Name:  "config, c",
-			Usage: "Tool configuration file (YAML)",
-		},
-	}
-	app.Before = func(c *cli.Context) error {
-		config.Default()
-
-		var err error
-
-		if c.String("config") != "" {
-			if err = config.LoadFromFile(c.String("config")); err != nil {
-				fmt.Printf("error loading configuration file. details: %s\n", err)
-				return err
-			}
-		}
-
-		if err = config.LoadFromEnvironment(); err != nil {
-			fmt.Printf("error loading configuration from environment variables. details: %s\n", err)
-			return err
-		}
-
-		logger = logrus.New()
-		logger.Out = os.Stdout
-
-		// optionally set logger output file defined in configuration. if not
-		// defined stdout will be used
-		if config.Current().Log.File != "" {
-			if logFile, err = os.OpenFile(config.Current().Log.File, os.O_CREATE|os.O_WRONLY|os.O_APPEND, os.ModePerm); err != nil {
-				fmt.Printf("error opening log file “%s”. details: %s\n", config.Current().Log.File, err)
-				return err
-			}
-
-			// writes to the stdout and to the log file
-			logger.Out = io.MultiWriter(os.Stdout, logFile)
-		}
-
-		switch config.Current().Log.Level {
-		case config.LogLevelDebug:
-			logger.Level = logrus.DebugLevel
-		case config.LogLevelInfo:
-			logger.Level = logrus.InfoLevel
-		case config.LogLevelWarning:
-			logger.Level = logrus.WarnLevel
-		case config.LogLevelError:
-			logger.Level = logrus.ErrorLevel
-		case config.LogLevelFatal:
-			logger.Level = logrus.FatalLevel
-		case config.LogLevelPanic:
-			logger.Level = logrus.PanicLevel
-		}
-
-		awsConfig := cloud.AWSConfig{
-			AccountID:       config.Current().AWS.AccountID.Value,
-			AccessKeyID:     config.Current().AWS.AccessKeyID.Value,
-			SecretAccessKey: config.Current().AWS.SecretAccessKey.Value,
-			Region:          config.Current().AWS.Region,
-			VaultName:       config.Current().AWS.VaultName,
-		}
-
-		var awsCloud cloud.Cloud
-		if awsCloud, err = cloud.NewAWSCloud(logger, awsConfig, false); err != nil {
-			fmt.Printf("error initializing AWS cloud. details: %s\n", err)
-			return err
-		}
-
-		var localStorage storage.Storage
-		switch config.Current().Database.Type {
-		case config.DatabaseTypeAuditFile:
-			localStorage = storage.NewAuditFile(logger, config.Current().Database.File)
-		case config.DatabaseTypeBoltDB:
-			localStorage = storage.NewBoltDB(logger, config.Current().Database.File)
-		}
-
-		toGlacier = ToGlacier{
-			context: ctx,
-			builder: archive.NewTARBuilder(logger),
-			envelop: archive.NewOFBEnvelop(logger),
-			cloud:   awsCloud,
-			storage: localStorage,
-			logger:  logger,
-		}
-
-		return nil
-	}
-	app.Commands = []cli.Command{
-		{
-			Name:  "sync",
-			Usage: "backup now the desired paths to AWS Glacier",
-			Action: func(c *cli.Context) error {
-				if err := toGlacier.Backup(config.Current().Paths, config.Current().BackupSecret.Value); err != nil {
-					logger.Error(err)
-				}
-				return nil
-			},
-		},
-		{
-			Name:      "get",
-			Usage:     "retrieve a specific backup from AWS Glacier",
-			ArgsUsage: "<archiveID>",
-			Action: func(c *cli.Context) error {
-				if err := toGlacier.RetrieveBackup(c.Args().First(), config.Current().BackupSecret.Value); err != nil {
-					logger.Error(err)
-				} else {
-					fmt.Println("Backup recovered successfully")
-				}
-				return nil
-			},
-		},
-		{
-			Name:      "remove",
-			Aliases:   []string{"rm"},
-			Usage:     "remove a specific backup from AWS Glacier",
-			ArgsUsage: "<archiveID>",
-			Action: func(c *cli.Context) error {
-				if err := toGlacier.RemoveBackup(c.Args().First()); err != nil {
-					logger.Error(err)
-				}
-				return nil
-			},
-		},
-		{
-			Name:    "list",
-			Aliases: []string{"ls"},
-			Usage:   "list all backups sent to AWS Glacier",
-			Flags: []cli.Flag{
-				cli.BoolFlag{
-					Name:  "remote,r",
-					Usage: "retrieve the list from AWS Glacier (long wait)",
-				},
-			},
-			Action: func(c *cli.Context) error {
-				if backups, err := toGlacier.ListBackups(c.Bool("remote")); err != nil {
-					logger.Error(err)
-
-				} else if len(backups) > 0 {
-					fmt.Println("Date             | Vault Name       | Archive ID")
-					fmt.Printf("%s-+-%s-+-%s\n", strings.Repeat("-", 16), strings.Repeat("-", 16), strings.Repeat("-", 138))
-					for _, backup := range backups {
-						fmt.Printf("%-16s | %-16s | %-138s\n", backup.Backup.CreatedAt.Format("2006-01-02 15:04"), backup.Backup.VaultName, backup.Backup.ID)
-					}
-				}
-				return nil
-			},
-		},
-		{
-			Name:  "start",
-			Usage: "run the scheduler (will block forever)",
-			Action: func(c *cli.Context) error {
-				scheduler := gocron.NewScheduler()
-				scheduler.Every(1).Day().At("00:00").Do(func() {
-					if err := toGlacier.Backup(config.Current().Paths, config.Current().BackupSecret.Value); err != nil {
-						logger.Error(err)
-					}
-				})
-				scheduler.Every(1).Weeks().At("01:00").Do(func() {
-					if err := toGlacier.RemoveOldBackups(config.Current().KeepBackups); err != nil {
-						logger.Error(err)
-					}
-				})
-				scheduler.Every(4).Weeks().At("12:00").Do(func() {
-					if _, err := toGlacier.ListBackups(true); err != nil {
-						logger.Error(err)
-					}
-				})
-				scheduler.Every(1).Weeks().At("06:00").Do(func() {
-					emailInfo := EmailInfo{
-						Sender:   EmailSenderFunc(smtp.SendMail),
-						Server:   config.Current().Email.Server,
-						Port:     config.Current().Email.Port,
-						Username: config.Current().Email.Username,
-						Password: config.Current().Email.Password.Value,
-						From:     config.Current().Email.From,
-						To:       config.Current().Email.To,
-					}
-
-					if err := toGlacier.SendReport(emailInfo); err != nil {
-						logger.Error(err)
-					}
-				})
-
-				stopped := scheduler.Start()
-				cancelFunc = func() {
-					close(stopped)
-				}
-
-				select {
-				case <-stopped:
-					// wait a small period just to give time for the scheduler to shutdown
-					time.Sleep(time.Second)
-				}
-
-				return nil
-			},
-		},
-		{
-			Name:  "report",
-			Usage: "test report notification",
-			Action: func(c *cli.Context) error {
-				report.Add(report.NewTest())
-
-				emailInfo := EmailInfo{
-					Sender:   EmailSenderFunc(smtp.SendMail),
-					Server:   config.Current().Email.Server,
-					Port:     config.Current().Email.Port,
-					Username: config.Current().Email.Username,
-					Password: config.Current().Email.Password.Value,
-					From:     config.Current().Email.From,
-					To:       config.Current().Email.To,
-				}
-
-				if err := toGlacier.SendReport(emailInfo); err != nil {
-					logger.Error(err)
-				}
-
-				return nil
-			},
-		},
-		{
-			Name:      "encrypt",
-			Aliases:   []string{"enc"},
-			Usage:     "encrypt a password or secret",
-			ArgsUsage: "<password>",
-			Action: func(c *cli.Context) error {
-				if pwd, err := config.PasswordEncrypt(c.Args().First()); err != nil {
-					logger.Error(err)
-				} else {
-					fmt.Printf("encrypted:%s\n", pwd)
-				}
-				return nil
-			},
-		},
-	}
-
-	// create a graceful shutdown when receiving a signal (SIGINT, SIGKILL,
-	// SIGTERM, SIGSTOP)
-	sigs := make(chan os.Signal, 1)
-	signal.Notify(sigs, syscall.SIGINT, syscall.SIGKILL, syscall.SIGTERM, syscall.SIGSTOP)
-
-	go func() {
-		<-sigs
-		if cancelFunc != nil {
-			cancelFunc()
-		}
-
-		cancel()
-	}()
-
-	app.Run(os.Args)
-}
 
 // ToGlacier manages backups in the cloud.
 type ToGlacier struct {
-	context context.Context
-	builder archive.Builder
-	envelop archive.Envelop
-	cloud   cloud.Cloud
-	storage storage.Storage
-	logger  log.Logger
+	Context context.Context
+	Builder archive.Builder
+	Envelop archive.Envelop
+	Cloud   cloud.Cloud
+	Storage storage.Storage
+	Logger  log.Logger
 }
 
 // Backup create an archive and send it to the cloud.
@@ -328,7 +47,7 @@ func (t ToGlacier) Backup(backupPaths []string, backupSecret string) error {
 	}
 
 	timeMark := time.Now()
-	filename, archiveInfo, err := t.builder.Build(archiveInfo, backupPaths...)
+	filename, archiveInfo, err := t.Builder.Build(archiveInfo, backupPaths...)
 	if err != nil {
 		backupReport.Errors = append(backupReport.Errors, err)
 		return errors.WithStack(err)
@@ -348,7 +67,7 @@ func (t ToGlacier) Backup(backupPaths []string, backupSecret string) error {
 		var encryptedFilename string
 
 		timeMark = time.Now()
-		if encryptedFilename, err = t.envelop.Encrypt(filename, backupSecret); err != nil {
+		if encryptedFilename, err = t.Envelop.Encrypt(filename, backupSecret); err != nil {
 			backupReport.Errors = append(backupReport.Errors, err)
 			return errors.WithStack(err)
 		}
@@ -361,7 +80,7 @@ func (t ToGlacier) Backup(backupPaths []string, backupSecret string) error {
 	}
 
 	timeMark = time.Now()
-	if backupReport.Backup, err = t.cloud.Send(t.context, filename); err != nil {
+	if backupReport.Backup, err = t.Cloud.Send(t.Context, filename); err != nil {
 		backupReport.Errors = append(backupReport.Errors, err)
 		return errors.WithStack(err)
 	}
@@ -375,7 +94,7 @@ func (t ToGlacier) Backup(backupPaths []string, backupSecret string) error {
 		}
 	}
 
-	if err := t.storage.Save(storage.Backup{Backup: backupReport.Backup, Info: archiveInfo}); err != nil {
+	if err := t.Storage.Save(storage.Backup{Backup: backupReport.Backup, Info: archiveInfo}); err != nil {
 		backupReport.Errors = append(backupReport.Errors, err)
 		return errors.WithStack(err)
 	}
@@ -390,7 +109,7 @@ func (t ToGlacier) ListBackups(remote bool) (storage.Backups, error) {
 		return t.listRemoteBackups()
 	}
 
-	backups, err := t.storage.List()
+	backups, err := t.Storage.List()
 	if err != nil {
 		return nil, errors.WithStack(err)
 	}
@@ -406,7 +125,7 @@ func (t ToGlacier) listRemoteBackups() (storage.Backups, error) {
 	}()
 
 	timeMark := time.Now()
-	remoteBackups, err := t.cloud.List(t.context)
+	remoteBackups, err := t.Cloud.List(t.Context)
 	if err != nil {
 		listBackupsReport.Errors = append(listBackupsReport.Errors, err)
 		return nil, errors.WithStack(err)
@@ -417,7 +136,7 @@ func (t ToGlacier) listRemoteBackups() (storage.Backups, error) {
 	// remote backups operations can take a while, and a concurrent action could
 	// change the local backups during this time
 
-	backups, err := t.storage.List()
+	backups, err := t.Storage.List()
 	if err != nil {
 		listBackupsReport.Errors = append(listBackupsReport.Errors, err)
 		return nil, errors.WithStack(err)
@@ -453,11 +172,11 @@ func (t ToGlacier) listRemoteBackups() (storage.Backups, error) {
 		if backup.Backup.CreatedAt.After(time.Now().Add(-24 * time.Hour)) {
 			// recent backups could not be in the inventory yet
 			kept = append(kept, backup.Backup.ID)
-			t.logger.Debugf("toglacier: backup id “%s” kept because is to recent", backup.Backup.ID)
+			t.Logger.Debugf("toglacier: backup id “%s” kept because is to recent", backup.Backup.ID)
 			continue
 		}
 
-		if err := t.storage.Remove(backup.Backup.ID); err != nil {
+		if err := t.Storage.Remove(backup.Backup.ID); err != nil {
 			listBackupsReport.Errors = append(listBackupsReport.Errors, err)
 			return nil, errors.WithStack(err)
 		}
@@ -469,12 +188,12 @@ func (t ToGlacier) listRemoteBackups() (storage.Backups, error) {
 	for i, remoteBackup := range remoteBackups {
 		// check if a recent backup appeared in the inventory
 		if j := sort.SearchStrings(kept, remoteBackup.ID); j < len(kept) && kept[j] == remoteBackup.ID {
-			if err := t.storage.Remove(kept[j]); err != nil {
+			if err := t.Storage.Remove(kept[j]); err != nil {
 				listBackupsReport.Errors = append(listBackupsReport.Errors, err)
 				return nil, errors.WithStack(err)
 			}
 
-			t.logger.Debugf("toglacier: backup id “%s” removed because it was found remotely", kept[j])
+			t.Logger.Debugf("toglacier: backup id “%s” removed because it was found remotely", kept[j])
 			kept = append(kept[:j], kept[j+1:]...)
 		}
 
@@ -496,7 +215,7 @@ func (t ToGlacier) listRemoteBackups() (storage.Backups, error) {
 			Info:   archiveInfo,
 		})
 
-		if err := t.storage.Save(syncBackups[i]); err != nil {
+		if err := t.Storage.Save(syncBackups[i]); err != nil {
 			listBackupsReport.Errors = append(listBackupsReport.Errors, err)
 			return nil, errors.WithStack(err)
 		}
@@ -519,7 +238,7 @@ func (t ToGlacier) listRemoteBackups() (storage.Backups, error) {
 
 // RetrieveBackup recover a specific backup from the cloud.
 func (t ToGlacier) RetrieveBackup(id, backupSecret string) error {
-	backups, err := t.storage.List()
+	backups, err := t.Storage.List()
 	if err != nil {
 		return errors.WithStack(err)
 	}
@@ -539,7 +258,7 @@ func (t ToGlacier) RetrieveBackup(id, backupSecret string) error {
 		// We will extract the archive information saved in the backup to detect all
 		// other backup parts that we need. This is important when the local storage
 		// got corrupted due to a disaster
-		filenames, err := t.cloud.Get(t.context, id)
+		filenames, err := t.Cloud.Get(t.Context, id)
 		if err != nil {
 			return errors.WithStack(err)
 		}
@@ -569,7 +288,7 @@ func (t ToGlacier) RetrieveBackup(id, backupSecret string) error {
 		ids = append(ids, id)
 	}
 
-	filenames, err := t.cloud.Get(t.context, ids...)
+	filenames, err := t.Cloud.Get(t.Context, ids...)
 	if err != nil {
 		return errors.WithStack(err)
 	}
@@ -590,7 +309,7 @@ func (t ToGlacier) decryptAndExtract(backupSecret, filename string, filter []str
 	if backupSecret != "" {
 		var decryptedFilename string
 
-		if decryptedFilename, err = t.envelop.Decrypt(filename, backupSecret); err != nil {
+		if decryptedFilename, err = t.Envelop.Decrypt(filename, backupSecret); err != nil {
 			return nil, errors.WithStack(err)
 		}
 
@@ -599,7 +318,7 @@ func (t ToGlacier) decryptAndExtract(backupSecret, filename string, filter []str
 		}
 	}
 
-	archiveInfo, err := t.builder.Extract(filename, filter)
+	archiveInfo, err := t.Builder.Extract(filename, filter)
 	if err != nil {
 		return nil, errors.WithStack(err)
 	}
@@ -611,11 +330,11 @@ func (t ToGlacier) decryptAndExtract(backupSecret, filename string, filter []str
 
 // RemoveBackup delete a specific backup from the cloud.
 func (t ToGlacier) RemoveBackup(id string) error {
-	if err := t.cloud.Remove(t.context, id); err != nil {
+	if err := t.Cloud.Remove(t.Context, id); err != nil {
 		return errors.WithStack(err)
 	}
 
-	return errors.WithStack(t.storage.Remove(id))
+	return errors.WithStack(t.Storage.Remove(id))
 }
 
 // RemoveOldBackups delete old backups from the cloud. This will optimize the
